auth-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
